+++ attribute/key.go
@@ -117,10 +117,10 @@ func (k Key) StringSlice(v []string) KeyValue {
 	}
 }
 
-// ByteSlice creates a KeyValue instance with a BYTES Value.
+// ByteSlice creates a KeyValue instance with a BYTESLICE Value.
 //
 // If creating both a key and value at the same time, use the provided
-// convenience function instead -- Bytes(name, value).
+// convenience function instead -- ByteSlice(name, value).
 func (k Key) ByteSlice(v []byte) KeyValue {
 	return KeyValue{
 		Key:   k,