+++ attribute/key_test.go
@@ -109,7 +109,7 @@ func TestEmit(t *testing.T) {
 			want: `["foo","bar"]`,
 		},
 		{
-			name: `test Key.Emit() can emit a string representing self.BYTES`,
+			name: `test Key.Emit() can emit a string representing self.BYTESLICE`,
 			v:    attribute.ByteSliceValue([]byte("foo")),
 			want: "Zm9v",
 		},